Move the get_logged_in_user to the common functions
